Add get and patch routes for proxy to stock-chart
